EDUARDO - Modelagem de todas as classes feitas, comentários onde pode haver ou terá alterações

+++ main/src/Avaliacao.java
@@ -3,58 +3,75 @@ public class Avaliacao {
     //Atributos
 
     //PK da tabela
-    private int id;
+    private long id;
 
     //FK originada da tabela perfil
-    private int idPerfil;
+    private long idPerfil;
 
     private String comentario;
     private int quantidadeEstrelas;
 
     //Construtor
-    public Avaliacao(int id, int idPerfil, String comentario, int quantidadeEstrelas) {
+    public Avaliacao(long id, long idPerfil, String comentario, int quantidadeEstrelas) {
+
         this.id = id;
         this.idPerfil = idPerfil;
         this.comentario = comentario;
         this.quantidadeEstrelas = quantidadeEstrelas;
+
     }
 
     //Getters e Setters
 
     //OBS: O atributo id não tem setter, pois ele é a pk da tabela
-    public int getId() {
+    public long getId() {
+
         return id;
+
     }
 
-    //OBS: O atributo idPerfil não tem setter, pois ele é a fk da tabela e nesse caso ela acaba sendo inválida
-    public int getIdPerfil() {
+    //OBS: O atributo idPerfil é uma FK e é imutável, então não possui setter
+    public long getIdPerfil() {
+
         return idPerfil;
+
     }
 
     public String getComentario() {
+
         return comentario;
+
     }
 
     public void setComentario(String comentario) {
+
         this.comentario = comentario;
+
     }
 
     public int getQuantidadeEstrelas() {
+
         return quantidadeEstrelas;
+
     }
 
     public void setQuantidadeEstrelas(int quantidadeEstrelas) {
+
         this.quantidadeEstrelas = quantidadeEstrelas;
+
     }
 
     //Método toString
     @Override
     public String toString(){
+
         return  "=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=\n" +
                 "ID: "+ this.id + "\n" +
                 "ID do perfil: "+ this.idPerfil + "\n" +
                 "Comentário: "+ this.comentario + "\n" +
                 "Quantidade de estrelas: "+ this.quantidadeEstrelas + "\n" +
                 "=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=\n";
+
     }
+
 }
