+++ main/src/Fornecedor.java
@@ -3,61 +3,76 @@ public class Fornecedor {
     //Atributos
 
     //PK da tabela
-    private int id;
+    private long id;
 
     //FK originada da tabela usuario
-    private int idUsuario;
+    private long idUsuario;
 
-    //Verificar esse tipo fornecedor
     private String tipoFornecedor;
-
     private String cnpj;
 
     //Construtor
-    public Fornecedor(int id, int idUsuario, String tipoFornecedor, String cnpj) {
+    public Fornecedor(long id, long idUsuario, String tipoFornecedor, String cnpj) {
+
         this.id = id;
         this.idUsuario = idUsuario;
         this.tipoFornecedor = tipoFornecedor;
         this.cnpj = cnpj;
+
     }
 
     //Getters e Setters
 
     //OBS: O atributo id não tem setter, pois ele é a pk da tabela
-    public int getId() {
+    public long getId() {
+
         return id;
+
     }
 
-    //OBS: O atributo idUsuario não tem setter, pois ele é a fk da tabela e nesse caso ela acaba sendo inválida
-    public int getIdUsuario() {
+    //OBS: O atributo idUsuario é uma FK e é imutável, então não possui setter
+    public long getIdUsuario() {
+
         return idUsuario;
+
     }
 
     public String getTipoFornecedor() {
+
         return tipoFornecedor;
+
     }
 
     public void setTipoFornecedor(String tipoFornecedor) {
+
         this.tipoFornecedor = tipoFornecedor;
+
     }
 
-    //Verificar se o cnpj vai possuir setter
+    //OBS: O atributo cnpj não tem setter, pois ele será imutável
     public String getCnpj() {
+
         return cnpj;
+
     }
 
     public void setCnpj(String cnpj) {
+
         this.cnpj = cnpj;
+
     }
 
     //Método toString
     @Override
     public String toString(){
+
         return  "=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=\n" +
                 "ID: "+ this.id + "\n" +
                 "ID do usuário: "+ this.idUsuario + "\n" +
                 "Tipo do fornecedor: "+ this.tipoFornecedor + "\n" +
                 "CNPJ: "+ this.cnpj + "\n" +
                 "=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=\n";
+
     }
+
 }
